requirements.txt
torch>=2.0.0
numpy>=1.24.0
gymnasium>=0.29.0
hydra-core>=1.3.0
omegaconf>=2.3.0
wandb[media]>=0.16.0
matplotlib>=3.7.0
pygame>=2.0.0
pytest>=7.0.0
